1. Remove useless shared method 'TrussUnitToolBar::addHideButton'. 2. Unshare 'AggToolBar::getButtons'. Now `protected'. 3. Tweaks to avoid gcc warnings. 4. Cosmetics.
git-svn-id: https://ferma.svn.sf.net/svnroot/ferma/trunk/FermaNext@307 d51c77f7-620d-0410-a109-9a2f07eaa567

+++ src/TrussUnit/TrussUnit.cpp
@@ -166,8 +166,8 @@ void TrussUnit::paintLoad ( TrussLoad& load, QPoint tailPos, ren_dynarow& baseRe
     else if ( y > 0 )
         forceDirection.setY ( 1 );
 
-    QPoint headPos ( tailPos.x() + 22 * x1 * forceDirection.x(), 
-                     tailPos.y() - 22 * y1 * forceDirection.y() );
+    QPoint headPos( int( tailPos.x() + 22 * x1 * forceDirection.x() ), 
+                    int( tailPos.y() - 22 * y1 * forceDirection.y() ) );
     
     solidRenderer solidRend ( baseRend );
     scanline_rasterizer   ras;
